moved kup to root and fixed formatting/lint issues
