+++ kup/.cruft.json
@@ -1,0 +1,18 @@
+{
+  "template": "https://github.com/runtimeverification/python-project-template.git",
+  "commit": "19e10414b54792af5aabf288bc2984fbcde44be8",
+  "checkout": null,
+  "context": {
+    "cookiecutter": {
+      "project_name": "kup",
+      "project_slug": "K Framework installer",
+      "package_name": "kup",
+      "version": "0.1.0",
+      "description": "",
+      "author_name": "Runtime Verification, Inc.",
+      "author_email": "contact@runtimeverification.com",
+      "_template": "https://github.com/runtimeverification/python-project-template.git"
+    }
+  },
+  "directory": null
+}
